requirements.txt
asdl==0.1.5
attrs==19.3.0
boto==2.49.0
boto3==1.14.44
botocore==1.17.44
bpemb==0.3.2
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
dill==0.3.1.1
docutils==0.15.2
entmax==1.0
future==0.18.2
gensim==3.8.3
jsonnet==0.16.0
networkx==2.4
nltk==3.5
numpy==1.21.4
pandas==1.1.3
pillow==8.0.1
protobuf==3.13.0
psutil==5.7.2
pyparsing==2.4.7
pyrsistent==0.16.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
spacy==2.3.2
stanford-corenlp==3.9.2
stanfordcorenlp==3.9.1.1
stanza==1.1.1
thinc==7.4.1
tokenizers==0.8.1rc1 
torchtext==0.8.1
tqdm==4.48.2
transformers==3.0.2
